# swagger-ui/dist/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Flutter Egypt Mobile App Server",
    "title": "Flutter Egypt API",
    "contact": {
      "name": "Flutter Egypt",
      "url": "http://www.facebook.com/FlutterEgypt",
      "email": "FlutterEgypt@google.com"
    },

    "license": {
      "name": "Author: Karim Mohamed",
      "url": "https://github.com/KarimMohamed2005"
    },
    "version": "1.0"
  },
  "tags": [
    {
      "name":"Users",
      "description": "All users operations"
    },
    {
      "name": "Posts",
      "description": "All posts operations"
    }
  ],
  "paths": {
    "/login": {
      "post": {

        "tags": [
          "Users"
        ],

        "parameters": [
          {
            "type": "string",
            "name": "email",
            "required": true,
            "in": "body"
          },
          {
            "type":"string",
            "name":"password",
            "required": true,
            "in": "body"
          }

        ],
        "responses": {
          "200": {
            "description": "UID, Email and Token",
            "headers": {
              "id": {
                "type": "string",
                "description": "UID (USER ID)"
              },
              "email": {
                "type":"string",
                "description": "USER EMAIL"
              },
              "token": {
                "type":"string",
                "description": "USER ACCESS TOKEN"
              }
            },
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/register": {
      "post": {

        "tags": [
          "Users"
        ],

        "parameters": [
          {
            "type": "string",
            "name": "email",
            "required": true,
            "in": "body"
          },
          {
            "type": "string",
            "name":"password",
            "required": true,
            "in": "body"
          }
        ],
        "responses": {
          "200": {
            "description": "Register Response",
            "headers": {
              "id": {
                "type": "string",
                "description": "UID (USER ID)"
              },
              "email": {
                "type": "string",
                "description": "USER EMAIL"
              }
            },
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/posts": {
      "get": {

        "tags": [
          "Posts"
        ],

        "parameters": [
          {
            "name": "Authorization",
            "type": "string",
            "required": true,
            "in": "header",
            "description": "Get Posts of user by token"
          }
        ],
        "responses": {
          "200": {
            "description": "Your Posts"
          }
        }
      }
    },
    "/posts": {
      "post": {

        "tags": [
          "Posts"
        ],

        "parameters": [
          {
            "type": "string",
            "name": "name",
            "required": true,
            "in": "body"
          },
          {
            "type": "string",
            "name":"title",
            "required": true,
            "in": "body"
          },
          {
            "type": "string",
            "name":"content",
            "required": true,
            "in": "body"
          }
        ],
        "responses": {
          "200": {
            "description": "Created Post"
          }
        }
      }
    }
  }
}